Introduce common fixtures frozen_time and mock_version
