+++ test/test_integration.py
@@ -1,20 +1,13 @@
-import datetime
 from unittest.mock import Mock
 
-import freezegun
 import yaml
 
 from security_constraints.main import main
 
 
-@freezegun.freeze_time(
-    time_to_freeze=datetime.datetime(
-        1986, 4, 9, 12, 11, 10, 9, tzinfo=datetime.timezone.utc
-    )
-)
-def test_main_flow(tmp_path, github_token, monkeypatch, requests_mock) -> None:
-    mock_version = Mock(return_value="x.y.z")
-    monkeypatch.setattr("security_constraints.main.version", mock_version)
+def test_main_flow(
+    frozen_time, tmp_path, github_token, monkeypatch, mock_version: Mock, requests_mock
+) -> None:
     output_file = str(tmp_path / "output")
     config_file = str(tmp_path / "sc-conf.yaml")
     ignored_ids = [3, 5]
